api_container/requirements.txt
python-dotenv
fastapi
uvicorn
requests
SQLAlchemy
